requirements.txt
requests==2.31.0
pandas==2.1.4
fake-useragent==1.4.0
streamlit==1.31.0
plotly==5.18.0
numpy==1.26.3
openai>=1.86.0
python-dotenv==1.0.1
langchain-core>=0.3.72
langchain-community>=0.3.27
langchain-openai>=0.3.28 